backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
pydantic==2.9.2
scikit-learn==1.5.2
python-dotenv==1.0.1
httpx==0.27.2
alembic==1.16.2
pandas==2.2.3
joblib==1.4.2
razorpay==2.0.1
